requirements.txt
streamlit
pandas
python-dotenv
openpyxl
langchain
langchain-core 
langchain-openai 